+++ src/game-ui/Canvas/CanvasCell.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Circle, Group, Image, Rect, Text } from 'react-konva';
 
 import { CELL_WIDTH_IN_PIXELS } from '../../constants/config';
+import { CREATURES } from '../../constants/creatures';
 import { getItem } from '../../constants/items';
 import {
   DEFAULT_FONT_COLOR,
@@ -10,7 +11,7 @@ import {
   TileType,
 } from '../../constants/tiles';
 import { GameAction, gameActions, HoverCellPayload } from '../../game-logic/game';
-import { CellContent } from '../../typings/cell';
+import { CellContent, CreatureData } from '../../typings/cell';
 import { Position } from '../../typings/position';
 import { Visibility } from '../../typings/visibility';
 import { Flame } from './Flame';
@@ -24,7 +25,9 @@ export interface CellProps {
   position: Position;
   itemsImage: HTMLImageElement | undefined;
   flameImage: HTMLImageElement | undefined;
+  creaturesImage: HTMLImageElement | undefined;
   burning: boolean;
+  creature?: CreatureData;
 }
 
 export const CanvasCell: React.FC<CellProps> = ({
@@ -37,6 +40,8 @@ export const CanvasCell: React.FC<CellProps> = ({
   itemsImage,
   flameImage,
   burning,
+  creature,
+  creaturesImage,
 }) => {
   const item = content !== 0 && content !== 'Player' ? getItem(content) : '';
   const tile = getTile(tileType);
@@ -59,6 +64,7 @@ export const CanvasCell: React.FC<CellProps> = ({
   };
 
   const Item = () => {
+    // TODO DRY with Creature
     if (content !== 0 && content !== 'Player') {
       if (visibility === 'dark' && !revealed) {
         return null;
@@ -101,6 +107,52 @@ export const CanvasCell: React.FC<CellProps> = ({
     return null;
   };
 
+  const Creature = () => {
+    // TODO DRY with Item
+    // Unlike items, we don't "remember" creatures
+    if (creature) {
+      if (visibility === 'dark') {
+        return null;
+      }
+      if (visibility === 'dim') {
+        return (
+          <Circle
+            x={position[0] * CELL_WIDTH_IN_PIXELS + CELL_WIDTH_IN_PIXELS / 2}
+            y={position[1] * CELL_WIDTH_IN_PIXELS + CELL_WIDTH_IN_PIXELS / 2}
+            radius={5}
+            fill={'orange'}
+            opacity={0.3}
+          />
+        );
+      }
+
+      let opacity = 0;
+      if (visibility === 'clear') {
+        opacity = 1;
+      } else if (visibility === 'dim') {
+        opacity = 0.3;
+      }
+      const spritePosition = CREATURES[creature.type].spritePosition;
+      return (
+        <Image
+          x={position[0] * CELL_WIDTH_IN_PIXELS + 1}
+          y={position[1] * CELL_WIDTH_IN_PIXELS + 1}
+          image={creaturesImage}
+          width={CELL_WIDTH_IN_PIXELS - 2}
+          height={CELL_WIDTH_IN_PIXELS - 2}
+          opacity={opacity}
+          crop={{
+            x: spritePosition[0] * 16,
+            y: spritePosition[1] * 16,
+            width: 16,
+            height: 16,
+          }}
+        />
+      );
+    }
+    return null;
+  };
+
   const renderTile = () => {
     if (visibility !== 'dark' || revealed) {
       return (
@@ -117,6 +169,10 @@ export const CanvasCell: React.FC<CellProps> = ({
   };
 
   const renderContentOrTile = () => {
+    if (visibility !== 'dark' && creature) {
+      return <Creature />;
+    }
+
     if (visibility !== 'dark' && burning) {
       return renderFlame();
     }
@@ -150,7 +206,9 @@ export const CanvasCell: React.FC<CellProps> = ({
 
   return (
     <Group
-      onMouseEnter={() => handleMouseEnter({ tileType, visibility, revealed, content, burning })}
+      onMouseEnter={() =>
+        handleMouseEnter({ tileType, visibility, revealed, content, burning, creature })
+      }
       onMouseLeave={() => handleMouseLeave()}
     >
       <Rect