+++ src/game-ui/MapGenerator/MapGenerator.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
+import seedrandom from 'seedrandom';
 import styled from 'styled-components';
 
 import { NUMBER_OF_ROUNDS_BURNING } from '../../constants/config';
+import { CreatureType } from '../../constants/creatures';
 import { TileType } from '../../constants/tiles';
 import { gameActions } from '../../game-logic/game';
 import { GameAction, GameState } from '../../game-logic/game';
-// import seedrandom from 'seedrandom';
 import { generateLevel } from '../../pcg/generateLevel';
 import { CellContent, CellData } from '../../typings/cell';
 import { Effect } from '../../typings/effect';
 import { Position } from '../../typings/position';
+import { getRandomString } from '../../utils/getRandomString';
 import Map from './Map';
 import { Toolbar } from './Toolbar';
 
@@ -29,15 +31,22 @@ interface Props {
 export const MapGenerator: React.FC<Props> = (props) => {
   const [selectedTile, setSelectedTile] = React.useState<TileType | null>(null);
   const [selectedContent, setSelectedContent] = React.useState<CellContent | null>(null);
+  const [selectedCreature, setSelectedCreature] = React.useState<CreatureType | null>(null);
   const [selectedEffect, setSelectedEffect] = React.useState<Effect | null>(null);
+  const [currentSeed, setCurrentSeed] = React.useState('');
+
+  const load = (seed: string) => {
+    const rng = seedrandom(seed);
+    const level = generateLevel(rng);
+    props.dispatch(gameActions.initPlayerSpawn(level.playerSpawn));
+    props.dispatch(gameActions.setSeed(seed));
+    props.dispatch(gameActions.setCurrentMap(level.gameMap));
+  };
 
   React.useEffect(() => {
-    // We can seed the rng by providing a string to seedrandom():
-    // const seed = seedrandom('hello');
-    // const newMap = generateMap(String(seed()));
-    const level = generateLevel();
-    props.dispatch(gameActions.initPlayerSpawn(level.playerSpawn));
-    props.dispatch(gameActions.setCurrentMap(level.gameMap));
+    const seed = getRandomString();
+    setCurrentSeed(seed);
+    load(seed);
   }, []);
 
   if (props.state.currentMap === null) {
@@ -47,6 +56,7 @@ export const MapGenerator: React.FC<Props> = (props) => {
   const resetSelected = () => {
     setSelectedTile(null);
     setSelectedContent(null);
+    setSelectedCreature(null);
     setSelectedEffect(null);
   };
 
@@ -58,6 +68,11 @@ export const MapGenerator: React.FC<Props> = (props) => {
   const handleSelectedContent = (content: CellContent) => {
     resetSelected();
     setSelectedContent(content);
+  };
+
+  const handleSelectedCreature = (creatureType: CreatureType) => {
+    resetSelected();
+    setSelectedCreature(creatureType);
   };
 
   const handleSelectedEffect = (effect: Effect) => {
@@ -72,7 +87,19 @@ export const MapGenerator: React.FC<Props> = (props) => {
     }
 
     if (selectedContent) {
-      updatedCellData.content = selectedContent;
+      if (updatedCellData.content) {
+        updatedCellData.content = 0;
+      } else {
+        updatedCellData.content = selectedContent;
+      }
+    }
+
+    if (selectedCreature) {
+      if (updatedCellData.creature) {
+        delete updatedCellData.creature;
+      } else {
+        updatedCellData.creature = { id: 'temp_id', type: selectedCreature };
+      }
     }
 
     if (selectedEffect) {
@@ -93,8 +120,13 @@ export const MapGenerator: React.FC<Props> = (props) => {
         handleSelectedTile={handleSelectedTile}
         selectedContent={selectedContent}
         handleSelectedContent={handleSelectedContent}
+        selectedCreature={selectedCreature}
+        handleSelectedCreature={handleSelectedCreature}
         selectedEffect={selectedEffect}
         handleSelectedEffect={handleSelectedEffect}
+        seed={currentSeed}
+        setSeed={setCurrentSeed}
+        load={() => load(currentSeed)}
       />
       <Map gameMap={props.state.currentMap} handleCellClick={handleCellClick} />
     </Wrapper>