+++ src/game-ui/MapGenerator/Toolbar.tsx
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 import { CELL_WIDTH_IN_PIXELS } from '../../constants/config';
+import { CREATURES, CreatureType } from '../../constants/creatures';
 import { ITEMS } from '../../constants/items';
 import { TILES, TileType } from '../../constants/tiles';
 import { CellContent } from '../../typings/cell';
@@ -50,14 +51,19 @@ interface Props {
   handleSelectedTile: (tile: TileType) => void;
   selectedContent: CellContent | null;
   handleSelectedContent: (content: CellContent) => void;
+  selectedCreature: CreatureType | null;
+  handleSelectedCreature: (content: CreatureType) => void;
   selectedEffect: Effect | null;
   handleSelectedEffect: (effect: Effect) => void;
+  seed: string;
+  setSeed: React.Dispatch<React.SetStateAction<string>>;
+  load: () => void;
 }
 
 export const Toolbar: React.FC<Props> = (props) => {
   return (
     <Wrapper>
-      <div style={{ height: '90%', overflowY: 'scroll' }}>
+      <div style={{ height: '80%', overflowY: 'scroll' }}>
         <p>Tiles</p>
         {TILES.map((tile) => {
           return (
@@ -91,6 +97,20 @@ export const Toolbar: React.FC<Props> = (props) => {
             </CellWrapper>
           );
         })}
+        <p>Creatures</p>
+        {Object.keys(CREATURES).map((creature) => {
+          const type = creature as CreatureType;
+          return (
+            <CellWrapper
+              key={type}
+              selected={props.selectedCreature === type}
+              onClick={() => props.handleSelectedCreature(type)}
+            >
+              <Cell {...defaultCellProps} tileType={'.'} creature={type} />
+              <Label>{type}</Label>
+            </CellWrapper>
+          );
+        })}
         <p>Effects</p>
         <CellWrapper
           selected={props.selectedEffect === 'burn'}
@@ -99,9 +119,17 @@ export const Toolbar: React.FC<Props> = (props) => {
           <Label>Burn</Label>
         </CellWrapper>
       </div>
-      <div style={{ height: '10%' }}>
+      <div style={{ height: '20%' }}>
         <p>----</p>
         <Link to="/">PLAY</Link>
+        <input
+          style={{ marginTop: 15 }}
+          value={props.seed}
+          onChange={(event) => props.setSeed(event.target.value)}
+        />
+        <button style={{ marginTop: 15 }} onClick={props.load}>
+          Load from seed
+        </button>
       </div>
     </Wrapper>
   );